+++ sickgear/metadata/mediabrowser.py
@@ -254,7 +254,7 @@ class MediaBrowserMetadata(generic.GenericMetadata):
         tv_node = etree.Element("Series")
 
         try:
-            show_info = t[int(show_obj.prodid)]
+            show_info = t.get_show(show_obj.prodid, language=show_obj.lang)
         except BaseTVinfoShownotfound as e:
             logger.log("Unable to find show with id %s on %s, skipping it" %
                        (show_obj.prodid, sickgear.TVInfoAPI(show_obj.tvid).name), logger.ERROR)
@@ -415,7 +415,7 @@ class MediaBrowserMetadata(generic.GenericMetadata):
 
             t = sickgear.TVInfoAPI(ep_obj.show_obj.tvid).setup(**tvinfo_config)
 
-            show_info = t[ep_obj.show_obj.prodid]
+            show_info = t.get_show(ep_obj.show_obj.prodid, language=ep_obj.show_obj.lang)
         except BaseTVinfoShownotfound as e:
             raise exceptions_helper.ShowNotFoundException(ex(e))
         except BaseTVinfoError as e: