+++ CHANGES.md
@@ -1,7 +1,7 @@
 ﻿### 3.28.0 (2023-xx-xx xx:xx:00 UTC)
 
 * Update html5lib 1.1 (f87487a) to 1.2-dev (3e500bb)
-* Update package resource API 63.2.0 (3ae44cd) to 67.3.2 (b9bf2ec)
+* Update package resource API 63.2.0 (3ae44cd) to 67.5.1 (f51eccd)
 * Update urllib3 1.26.13 (25fbd5f) to 1.26.14 (a06c05c)
 * Change remove calls to legacy py2 fix encoding function
 * Change requirements for pure py3
@@ -12,6 +12,7 @@
 [develop changelog]
 
 * Add logging around the restart/shutdown event
+* Update package resource API 63.2.0 (3ae44cd) to 67.3.2 (b9bf2ec)
 
 
 ### 3.27.11 (2023-03-06 23:40:00 UTC)